// web/generated/schema.ts
export interface FrontPageSection extends SanityDocument {
  _type: "frontPageSection"

  /**
   * Headline — `localeString`
   *
   * Limit 15 characters.
   */
  title?: LocaleString

  /**
   * Subheading — `localeString`
   *
   * Limit 35 characters.
   */
  caption?: LocaleString

  /**
   * Heading — `localeString`
   *
   * Limit 30 characters.
   */
  heading?: LocaleString

  /**
   * Body — `localeRichText`
   *
   * Limit 400 characters.
   */
  body?: LocaleRichText

  /**
   * Call to Action — `localeString`
   *
   * A call to action for the internal link (20 characters max).
   */
  cta?: LocaleString

  /**
   * Link — `string`
   *
   *
   */
  ctaLink?: string

  /**
   * Main image — `captionImage`
   *
   * Images should be jpeg of 1440px along the longest edge, 500-600k is best.
   */
  mainImage?: CaptionImage

  /**
   * Inset image — `captionImage`
   *
   * Images should be jpeg of 1440px along the longest edge, 500-600k is best.
   */
  subImage?: CaptionImage
}
// ...
export type BlockContent = Array<
  | SanityKeyed<SanityBlock>
  | SanityKeyed<{
      _type: "image"
      asset: SanityReference<SanityImageAsset>
      crop?: SanityImageCrop
      hotspot?: SanityImageHotspot
    }>
>

export type CaptionImage = {
  _type: "image"
  asset: SanityReference<SanityImageAsset>
  crop?: SanityImageCrop
  hotspot?: SanityImageHotspot

  /**
   * Caption — `string`
   *
   *
   */
  caption?: string

  /**
   * Attribution — `string`
   *
   *
   */
  Caption?: string
}

export type LocaleRichText = {
  _type: "localeRichText"
  /**
   * English — `blockContent`
   *
   *
   */
  en?: BlockContent

  /**
   * Welsh — `blockContent`
   *
   *
   */
  cy?: BlockContent
}

export type LocaleSlug = {
  _type: "localeSlug"
  /**
   * English — `slug`
   *
   *
   */
  en?: { _type: "en"; current: string }

  /**
   * Welsh — `slug`
   *
   *
   */
  cy?: { _type: "cy"; current: string }
}

export type LocaleString = {
  _type: "localeString"
  /**
   * English — `string`
   *
   *
   */
  en?: string

  /**
   * Welsh — `string`
   *
   *
   */
  cy?: string
}